refactor: moved controllers and routes to presentation layer
